books_database/requirements.txt
grpcio==1.71.0
grpcio-tools==1.71.0
protobuf==5.29.0
watchdog==6.0.0